@@ -15,145 +15,162 @@ use CodeIgniter\Config\BaseConfig;
  */
 class ContentSecurityPolicy extends BaseConfig
 {
-	//-------------------------------------------------------------------------
-	// Broadbrush CSP management
-	//-------------------------------------------------------------------------
+    // -------------------------------------------------------------------------
+    // Broadbrush CSP management
+    // -------------------------------------------------------------------------
 
-	/**
-	 * Default CSP report context
-	 *
-	 * @var boolean
-	 */
-	public $reportOnly = false;
+    /**
+     * Default CSP report context
+     */
+    public bool $reportOnly = false;
 
-	/**
-	 * Specifies a URL where a browser will send reports
-	 * when a content security policy is violated.
-	 *
-	 * @var string|null
-	 */
-	public $reportURI = null;
+    /**
+     * Specifies a URL where a browser will send reports
+     * when a content security policy is violated.
+     */
+    public ?string $reportURI = null;
 
-	/**
-	 * Instructs user agents to rewrite URL schemes, changing
-	 * HTTP to HTTPS. This directive is for websites with
-	 * large numbers of old URLs that need to be rewritten.
-	 *
-	 * @var boolean
-	 */
-	public $upgradeInsecureRequests = false;
+    /**
+     * Instructs user agents to rewrite URL schemes, changing
+     * HTTP to HTTPS. This directive is for websites with
+     * large numbers of old URLs that need to be rewritten.
+     */
+    public bool $upgradeInsecureRequests = false;
 
-	//-------------------------------------------------------------------------
-	// Sources allowed
-	// Note: once you set a policy to 'none', it cannot be further restricted
-	//-------------------------------------------------------------------------
+    // -------------------------------------------------------------------------
+    // Sources allowed
+    // Note: once you set a policy to 'none', it cannot be further restricted
+    // -------------------------------------------------------------------------
 
-	/**
-	 * Will default to self if not overridden
-	 *
-	 * @var string|string[]|null
-	 */
-	public $defaultSrc = null;
+    /**
+     * Will default to self if not overridden
+     *
+     * @var string|string[]|null
+     */
+    public $defaultSrc;
 
-	/**
-	 * Lists allowed scripts' URLs.
-	 *
-	 * @var string|string[]
-	 */
-	public $scriptSrc = 'self';
+    /**
+     * Lists allowed scripts' URLs.
+     *
+     * @var string|string[]
+     */
+    public $scriptSrc = 'self';
 
-	/**
-	 * Lists allowed stylesheets' URLs.
-	 *
-	 * @var string|string[]
-	 */
-	public $styleSrc = 'self';
+    /**
+     * Lists allowed stylesheets' URLs.
+     *
+     * @var string|string[]
+     */
+    public $styleSrc = 'self';
 
-	/**
-	 * Defines the origins from which images can be loaded.
-	 *
-	 * @var string|string[]
-	 */
-	public $imageSrc = 'self';
+    /**
+     * Defines the origins from which images can be loaded.
+     *
+     * @var string|string[]
+     */
+    public $imageSrc = 'self';
 
-	/**
-	 * Restricts the URLs that can appear in a page's `<base>` element.
-	 *
-	 * Will default to self if not overridden
-	 *
-	 * @var string|string[]|null
-	 */
-	public $baseURI = null;
+    /**
+     * Restricts the URLs that can appear in a page's `<base>` element.
+     *
+     * Will default to self if not overridden
+     *
+     * @var string|string[]|null
+     */
+    public $baseURI;
 
-	/**
-	 * Lists the URLs for workers and embedded frame contents
-	 *
-	 * @var string|string[]
-	 */
-	public $childSrc = 'self';
+    /**
+     * Lists the URLs for workers and embedded frame contents
+     *
+     * @var string|string[]
+     */
+    public $childSrc = 'self';
 
-	/**
-	 * Limits the origins that you can connect to (via XHR,
-	 * WebSockets, and EventSource).
-	 *
-	 * @var string|string[]
-	 */
-	public $connectSrc = 'self';
+    /**
+     * Limits the origins that you can connect to (via XHR,
+     * WebSockets, and EventSource).
+     *
+     * @var string|string[]
+     */
+    public $connectSrc = 'self';
 
-	/**
-	 * Specifies the origins that can serve web fonts.
-	 *
-	 * @var string|string[]
-	 */
-	public $fontSrc = null;
+    /**
+     * Specifies the origins that can serve web fonts.
+     *
+     * @var string|string[]
+     */
+    public $fontSrc;
 
-	/**
-	 * Lists valid endpoints for submission from `<form>` tags.
-	 *
-	 * @var string|string[]
-	 */
-	public $formAction = 'self';
+    /**
+     * Lists valid endpoints for submission from `<form>` tags.
+     *
+     * @var string|string[]
+     */
+    public $formAction = 'self';
 
-	/**
-	 * Specifies the sources that can embed the current page.
-	 * This directive applies to `<frame>`, `<iframe>`, `<embed>`,
-	 * and `<applet>` tags. This directive can't be used in
-	 * `<meta>` tags and applies only to non-HTML resources.
-	 *
-	 * @var string|string[]|null
-	 */
-	public $frameAncestors = null;
+    /**
+     * Specifies the sources that can embed the current page.
+     * This directive applies to `<frame>`, `<iframe>`, `<embed>`,
+     * and `<applet>` tags. This directive can't be used in
+     * `<meta>` tags and applies only to non-HTML resources.
+     *
+     * @var string|string[]|null
+     */
+    public $frameAncestors;
 
-	/**
-	 * Restricts the origins allowed to deliver video and audio.
-	 *
-	 * @var string|string[]|null
-	 */
-	public $mediaSrc = null;
+    /**
+     * The frame-src directive restricts the URLs which may
+     * be loaded into nested browsing contexts.
+     *
+     * @var array|string|null
+     */
+    public $frameSrc;
 
-	/**
-	 * Allows control over Flash and other plugins.
-	 *
-	 * @var string|string[]
-	 */
-	public $objectSrc = 'self';
+    /**
+     * Restricts the origins allowed to deliver video and audio.
+     *
+     * @var string|string[]|null
+     */
+    public $mediaSrc;
 
-	/**
-	 * @var string|string[]|null
-	 */
-	public $manifestSrc = null;
+    /**
+     * Allows control over Flash and other plugins.
+     *
+     * @var string|string[]
+     */
+    public $objectSrc = 'self';
 
-	/**
-	 * Limits the kinds of plugins a page may invoke.
-	 *
-	 * @var string|string[]|null
-	 */
-	public $pluginTypes = null;
+    /**
+     * @var string|string[]|null
+     */
+    public $manifestSrc;
 
-	/**
-	 * List of actions allowed.
-	 *
-	 * @var string|string[]|null
-	 */
-	public $sandbox = null;
+    /**
+     * Limits the kinds of plugins a page may invoke.
+     *
+     * @var string|string[]|null
+     */
+    public $pluginTypes;
+
+    /**
+     * List of actions allowed.
+     *
+     * @var string|string[]|null
+     */
+    public $sandbox;
+
+    /**
+     * Nonce tag for style
+     */
+    public string $styleNonceTag = '{csp-style-nonce}';
+
+    /**
+     * Nonce tag for script
+     */
+    public string $scriptNonceTag = '{csp-script-nonce}';
+
+    /**
+     * Replace nonce tag automatically
+     */
+    public bool $autoNonce = true;
 }
