@@ -12,56 +12,72 @@ use CodeIgniter\Config\BaseConfig;
  */
 class Encryption extends BaseConfig
 {
-	/**
-	 * --------------------------------------------------------------------------
-	 * Encryption Key Starter
-	 * --------------------------------------------------------------------------
-	 *
-	 * If you use the Encryption class you must set an encryption key (seed).
-	 * You need to ensure it is long enough for the cipher and mode you plan to use.
-	 * See the user guide for more info.
-	 *
-	 * @var string
-	 */
-	public $key = '';
+    /**
+     * --------------------------------------------------------------------------
+     * Encryption Key Starter
+     * --------------------------------------------------------------------------
+     *
+     * If you use the Encryption class you must set an encryption key (seed).
+     * You need to ensure it is long enough for the cipher and mode you plan to use.
+     * See the user guide for more info.
+     */
+    public string $key = '';
 
-	/**
-	 * --------------------------------------------------------------------------
-	 * Encryption Driver to Use
-	 * --------------------------------------------------------------------------
-	 *
-	 * One of the supported encryption drivers.
-	 *
-	 * Available drivers:
-	 * - OpenSSL
-	 * - Sodium
-	 *
-	 * @var string
-	 */
-	public $driver = 'OpenSSL';
+    /**
+     * --------------------------------------------------------------------------
+     * Encryption Driver to Use
+     * --------------------------------------------------------------------------
+     *
+     * One of the supported encryption drivers.
+     *
+     * Available drivers:
+     * - OpenSSL
+     * - Sodium
+     */
+    public string $driver = 'OpenSSL';
 
-	/**
-	 * --------------------------------------------------------------------------
-	 * SodiumHandler's Padding Length in Bytes
-	 * --------------------------------------------------------------------------
-	 *
-	 * This is the number of bytes that will be padded to the plaintext message
-	 * before it is encrypted. This value should be greater than zero.
-	 *
-	 * See the user guide for more information on padding.
-	 *
-	 * @var integer
-	 */
-	public $blockSize = 16;
+    /**
+     * --------------------------------------------------------------------------
+     * SodiumHandler's Padding Length in Bytes
+     * --------------------------------------------------------------------------
+     *
+     * This is the number of bytes that will be padded to the plaintext message
+     * before it is encrypted. This value should be greater than zero.
+     *
+     * See the user guide for more information on padding.
+     */
+    public int $blockSize = 16;
 
-	/**
-	 * --------------------------------------------------------------------------
-	 * Encryption digest
-	 * --------------------------------------------------------------------------
-	 *
-	 * HMAC digest to use, e.g. 'SHA512' or 'SHA256'. Default value is 'SHA512'.
-	 *
-	 * @var string
-	 */
-	public $digest = 'SHA512';
+    /**
+     * --------------------------------------------------------------------------
+     * Encryption digest
+     * --------------------------------------------------------------------------
+     *
+     * HMAC digest to use, e.g. 'SHA512' or 'SHA256'. Default value is 'SHA512'.
+     */
+    public string $digest = 'SHA512';
+
+    /**
+     * Whether the cipher-text should be raw. If set to false, then it will be base64 encoded.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to false for CI3 Encryption compatibility.
+     */
+    public bool $rawData = true;
+
+    /**
+     * Encryption key info.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to 'encryption' for CI3 Encryption compatibility.
+     */
+    public string $encryptKeyInfo = '';
+
+    /**
+     * Authentication key info.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to 'authentication' for CI3 Encryption compatibility.
+     */
+    public string $authKeyInfo = '';
 }
