@@ -3,46 +3,75 @@
 namespace Config;
 
 use CodeIgniter\Config\BaseConfig;
+use Psr\Log\LogLevel;
 
 /**
  * Setup how the exception handler works.
  */
 class Exceptions extends BaseConfig
 {
-	/**
-	 * --------------------------------------------------------------------------
-	 * LOG EXCEPTIONS?
-	 * --------------------------------------------------------------------------
-	 * If true, then exceptions will be logged
-	 * through Services::Log.
-	 *
-	 * Default: true
-	 *
-	 * @var boolean
-	 */
-	public $log = true;
+    /**
+     * --------------------------------------------------------------------------
+     * LOG EXCEPTIONS?
+     * --------------------------------------------------------------------------
+     * If true, then exceptions will be logged
+     * through Services::Log.
+     *
+     * Default: true
+     */
+    public bool $log = true;
 
-	/**
-	 * --------------------------------------------------------------------------
-	 * DO NOT LOG STATUS CODES
-	 * --------------------------------------------------------------------------
-	 * Any status codes here will NOT be logged if logging is turned on.
-	 * By default, only 404 (Page Not Found) exceptions are ignored.
-	 *
-	 * @var array
-	 */
-	public $ignoreCodes = [404];
+    /**
+     * --------------------------------------------------------------------------
+     * DO NOT LOG STATUS CODES
+     * --------------------------------------------------------------------------
+     * Any status codes here will NOT be logged if logging is turned on.
+     * By default, only 404 (Page Not Found) exceptions are ignored.
+     */
+    public array $ignoreCodes = [404];
 
-	/**
-	 * --------------------------------------------------------------------------
-	 * Error Views Path
-	 * --------------------------------------------------------------------------
-	 * This is the path to the directory that contains the 'cli' and 'html'
-	 * directories that hold the views used to generate errors.
-	 *
-	 * Default: APPPATH.'Views/errors'
-	 *
-	 * @var string
-	 */
-	public $errorViewPath = APPPATH . 'Views/errors';
+    /**
+     * --------------------------------------------------------------------------
+     * Error Views Path
+     * --------------------------------------------------------------------------
+     * This is the path to the directory that contains the 'cli' and 'html'
+     * directories that hold the views used to generate errors.
+     *
+     * Default: APPPATH.'Views/errors'
+     */
+    public string $errorViewPath = APPPATH . 'Views/errors';
+
+    /**
+     * --------------------------------------------------------------------------
+     * HIDE FROM DEBUG TRACE
+     * --------------------------------------------------------------------------
+     * Any data that you would like to hide from the debug trace.
+     * In order to specify 2 levels, use "/" to separate.
+     * ex. ['server', 'setup/password', 'secret_token']
+     */
+    public array $sensitiveDataInTrace = [];
+
+    /**
+     * --------------------------------------------------------------------------
+     * LOG DEPRECATIONS INSTEAD OF THROWING?
+     * --------------------------------------------------------------------------
+     * By default, CodeIgniter converts deprecations into exceptions. Also,
+     * starting in PHP 8.1 will cause a lot of deprecated usage warnings.
+     * Use this option to temporarily cease the warnings and instead log those.
+     * This option also works for user deprecations.
+     */
+    public bool $logDeprecations = true;
+
+    /**
+     * --------------------------------------------------------------------------
+     * LOG LEVEL THRESHOLD FOR DEPRECATIONS
+     * --------------------------------------------------------------------------
+     * If `$logDeprecations` is set to `true`, this sets the log level
+     * to which the deprecation will be logged. This should be one of the log
+     * levels recognized by PSR-3.
+     *
+     * The related `Config\Logger::$threshold` should be adjusted, if needed,
+     * to capture logging the deprecations.
+     */
+    public string $deprecationLogLevel = LogLevel::WARNING;
 }
