@@ -8,28 +8,24 @@ use CodeIgniter\Images\Handlers\ImageMagickHandler;
 
 class Images extends BaseConfig
 {
-	/**
-	 * Default handler used if no other handler is specified.
-	 *
-	 * @var string
-	 */
-	public $defaultHandler = 'gd';
+    /**
+     * Default handler used if no other handler is specified.
+     */
+    public string $defaultHandler = 'gd';
 
-	/**
-	 * The path to the image library.
-	 * Required for ImageMagick, GraphicsMagick, or NetPBM.
-	 *
-	 * @var string
-	 */
-	public $libraryPath = '/usr/local/bin/convert';
+    /**
+     * The path to the image library.
+     * Required for ImageMagick, GraphicsMagick, or NetPBM.
+     */
+    public string $libraryPath = '/usr/local/bin/convert';
 
-	/**
-	 * The available handler classes.
-	 *
-	 * @var array<string, string>
-	 */
-	public $handlers = [
-		'gd'      => GDHandler::class,
-		'imagick' => ImageMagickHandler::class,
-	];
+    /**
+     * The available handler classes.
+     *
+     * @var array<string, string>
+     */
+    public array $handlers = [
+        'gd'      => GDHandler::class,
+        'imagick' => ImageMagickHandler::class,
+    ];
 }
