@@ -3,7 +3,7 @@
 namespace Config;
 
 use CodeIgniter\Config\BaseConfig;
-use Kint\Renderer\Renderer;
+use Kint\Renderer\AbstractRenderer;
 
 /**
  * --------------------------------------------------------------------------
@@ -17,45 +17,35 @@ use Kint\Renderer\Renderer;
  */
 class Kint extends BaseConfig
 {
-	/*
-	|--------------------------------------------------------------------------
-	| Global Settings
-	|--------------------------------------------------------------------------
-	*/
+    /*
+    |--------------------------------------------------------------------------
+    | Global Settings
+    |--------------------------------------------------------------------------
+    */
 
-	public $plugins = null;
+    public $plugins;
+    public int $maxDepth           = 6;
+    public bool $displayCalledFrom = true;
+    public bool $expanded          = false;
 
-	public $maxDepth = 6;
+    /*
+    |--------------------------------------------------------------------------
+    | RichRenderer Settings
+    |--------------------------------------------------------------------------
+    */
+    public string $richTheme = 'aante-light.css';
+    public bool $richFolder  = false;
+    public int $richSort     = AbstractRenderer::SORT_FULL;
+    public $richObjectPlugins;
+    public $richTabPlugins;
 
-	public $displayCalledFrom = true;
-
-	public $expanded = false;
-
-	/*
-	|--------------------------------------------------------------------------
-	| RichRenderer Settings
-	|--------------------------------------------------------------------------
-	*/
-	public $richTheme = 'aante-light.css';
-
-	public $richFolder = false;
-
-	public $richSort = Renderer::SORT_FULL;
-
-	public $richObjectPlugins = null;
-
-	public $richTabPlugins = null;
-
-	/*
-	|--------------------------------------------------------------------------
-	| CLI Settings
-	|--------------------------------------------------------------------------
-	*/
-	public $cliColors = true;
-
-	public $cliForceUTF8 = false;
-
-	public $cliDetectWidth = true;
-
-	public $cliMinWidth = 40;
+    /*
+    |--------------------------------------------------------------------------
+    | CLI Settings
+    |--------------------------------------------------------------------------
+    */
+    public bool $cliColors      = true;
+    public bool $cliForceUTF8   = false;
+    public bool $cliDetectWidth = true;
+    public int $cliMinWidth     = 40;
 }
