@@ -57,7 +57,7 @@ abstract class GoBaseResourceController extends \CodeIgniter\RESTful\ResourceCon
      *
      * @var array
      */
-    protected $helpers = ['session', 'go_common', 'form'];
+    protected $helpers = ['session', 'go_common', 'form', 'text'];
 
     /**
      * Initializer method.
@@ -78,7 +78,7 @@ abstract class GoBaseResourceController extends \CodeIgniter\RESTful\ResourceCon
 
         if ($this->usePageSubTitle) {
             $this->pageSubTitle = config('Basics')->appName;
-            $this->viewData['pageSubTitle'] = ' in ' . $this->pageSubTitle;
+            $this->viewData['pageSubTitle'] = ' ' . $this->pageSubTitle;
         }
         $this->viewData['errorMessage'] = $this->session->getFlashdata('errorMessage');
         $this->viewData['successMessage'] = $this->session->getFlashdata('successMessage');
@@ -95,8 +95,16 @@ abstract class GoBaseResourceController extends \CodeIgniter\RESTful\ResourceCon
         $this->viewData['usingSweetAlert'] = true;
 
         $this->viewData['viewPath'] = static::$viewPath;
+        
+        $this->viewData['currentLocale'] = $this->request->getLocale();
     }
 
+    /**
+     * Convenience method to display the form of a module
+     * @param $forMethod
+     * @param null $objId
+     * @return string
+     */
     protected function displayForm($forMethod, $objId = null)
     {
 
@@ -131,7 +139,7 @@ abstract class GoBaseResourceController extends \CodeIgniter\RESTful\ResourceCon
 
         $viewFilePath = static::$viewPath . 'view' . ucfirst(static::$singularObjectNameCc) . 'Form';
 
-        echo view($viewFilePath, $this->viewData);
+        return view($viewFilePath, $this->viewData);
     }
 
     protected function redirect2listView($flashDataKey = null, $flashDataValue = null)
@@ -183,42 +191,137 @@ abstract class GoBaseResourceController extends \CodeIgniter\RESTful\ResourceCon
      */
     public function delete($id = null)
     {
-        if (!$this->model->delete($id)) {
-            return $this->failNotFound('No ' . static::$singularObjectName . ' could be deleted now, because it was probably deleted already or it did not exist.');
+        if (!empty(static::$pluralObjectNameCc) && !empty(static::$singularObjectNameCc)) {
+            $objName = mb_strtolower(lang(ucfirst(static::$pluralObjectNameCc).'.'.static::$singularObjectNameCc));
+        } else {
+            $objName = lang('Basic.global.record');
         }
-
-        return $this->respondDeleted(['id' => $id], 'The ' . static::$singularObjectName . ' was successfully deleted.');
+        
+        if (!$this->model->delete($id)) {
+            return $this->failNotFound(lang('Basic.global.deleteError', [$objName]));
+        }
+        
+        $message = lang('Basic.global.deleteSuccess', [$objName]);
+        $response = $this->respondDeleted(['id' => $id, 'msg' => $message]);
+        return $response;
     }
 
 
-    
-
+    /**
+     * Convenience method to validate form submission
+     * @return bool
+     */
     protected function canValidate()
     {
 
-        $validationRules = $this->formValidationRules ?? null;
+        $validationRules = $this->model->validationRules ?? $this->formValidationRules ?? null;
 
         if ($validationRules == null) {
             return true;
         }
 
-        $validationErrors = $this->formValidationErrors ?? null;
+        $validationErrorMessages = $this->model->validationMessages ??  $this->formValidationErrorMessagess ?? null;;
 
-        $validation = \Config\Services::validation();
-
-        // $validation->setRules($validationRules, $validationErrors)
-
-        if ($validationErrors != null) {
-            $valid = $this->validate($validationRules, $validationErrors);
+        if ($validationErrorMessages != null) {
+            $valid = $this->validate($validationRules, $validationErrorMessages);
         } else {
             $valid = $this->validate($validationRules);
         }
+        
+        $this->validationErrors = $valid ? '' : $this->validator->getErrors();
 
-        /* // As of version 1.1.5 of CodeIgniter Wizard, the following is replaced by custom validation errors template supported by CodeIgniter 4
+        /*
+        // As of version 1.1.5 of CodeIgniter Wizard, the following is replaced by custom validation errors template supported by CodeIgniter 4
+        // If you are not using Bootstrap templates, however, you might want to uncomment this block...
+        $validation =  \Config\Services::validation();
+        $this->validation = $validation;
         if (!$valid) {
             $this->viewData['errorMessage'] .= $validation->listErrors();
         }
         */
         return $valid;
+    }
+
+    /**
+     * Method for post(ed) input sanitization. Override this when you have custom transformation needs, etc.
+     * @param array|null $postData
+     * @return array
+     */
+    protected function sanitized(array $postData = null, bool $nullIfEmpty = false) {
+        if ($postData == null) {
+            $postData = $this->request->getPost();
+        }
+        $sanitizedData = [];
+        foreach ($postData as $k => $v) {
+            if ($k == csrf_token()) {
+                continue;
+            }
+            $sanitizationResult = goSanitize($v, $nullIfEmpty);
+            $sanitizedData[$k] = $sanitizationResult[0];
+        }
+        return $sanitizedData;
+    }
+    
+    /**
+     * Custom fail method needed when CSRF token regeneration is on in security settings
+     * @param string|array $messages
+     * @param int $status
+     * @param string|null $code
+     * @param string $customMessage
+     * @return mixed
+     */
+    protected function failWithNewToken($messages, int $status = 400, string $code = null, string $customMessage = '') {
+
+        if (! is_array($messages))
+        {
+            $messages = ['error' => $messages];
+        }
+        $response = [
+            'status'   => $status,
+            'error'    => $status,
+            'messages' => $messages,
+            csrf_token() => csrf_hash()
+        ];
+
+        return $this->respond($response, $status);
+    }
+
+    /**
+     * Used when a specified resource cannot be found and send back a new CSRF token.
+     *
+     * @param string $description
+     * @param string $code
+     * @param string $message
+     *
+     * @return mixed
+     */
+    public function failNotFoundWithNewToken(string $description = 'Not Found', string $code = null, string $message = '')
+    {
+        return $this->failWithNewToken($description, $this->codes['resource_not_found'], $code, $message);
+    }
+
+    /**
+     * Convenience method for common exception handling
+     * @param \Exception $e
+     */
+    protected function dealWithException(\Exception $e) {
+        // using another try / catch block to prevent to avoid CodeIgniter bug throwing trivial exceptions for querying DB errors
+        try {
+            $query = $this->model->db->getLastQuery();
+            $queryStr = !is_null($query) ? $query->getQuery() : '';
+            $dbError = $this->model->db->error();
+            $userFriendlyErrMsg = lang('Basic.global.persistErr1', [static::$singularObjectNameCc]);
+            if (isset($dbError['code']) && $dbError['code'] == 1062) :
+                $userFriendlyErrMsg .= PHP_EOL.lang('Basic.global.persistDuplErr', [static::$singularObjectNameCc]);
+            endif;
+            // $userFriendlyErrMsg = str_replace("'", "\'", $userFriendlyErrMsg); // Uncomment if experiencing unescaped single quote errors
+            log_message('error', $userFriendlyErrMsg.PHP_EOL.$e->getMessage().PHP_EOL.$queryStr);
+            if (isset($dbError['message']) && !empty($dbError['message'])) :
+                log_message('error', $dbError['code'].' : '.$dbError['message']);
+            endif;
+            $this->viewData['errorMessage'] = $userFriendlyErrMsg;
+        } catch (\Exception $e2) {
+            log_message('debug', 'You can probably safely ignore this: In attempt to check DB errors, CodeIgniter threw: '.PHP_EOL.$e2->getMessage());
+        }
     }
 }
