@@ -7,7 +7,7 @@ namespace App\Entities;
  *
  * @author Gökhan Ozar
  */
-abstract class GoBaseEntity extends \CodeIgniter\Entity
+abstract class GoBaseEntity extends \CodeIgniter\Entity\Entity
 {
     /**
      * Merges the current object's property values with the passed in arguments in array format
@@ -16,9 +16,11 @@ abstract class GoBaseEntity extends \CodeIgniter\Entity
      * @return $this - returns the current object instance
      */
     public function mergeAttributes(array $args = []) {
-        if (!empty($args)) :
-            $this->fill($args);
-        endif;
+        foreach ($args as $key => $value) {
+            if (isset($this->attributes[$key])) {
+                $this->attributes[$key] = $value;
+            }
+        }
         return $this;
     }
 }
