@@ -1,12 +1,9 @@
 		<li class="nav-item">
-			<?= anchor(route_to('countries'), '<i class="far fa-circle nav-icon"></i> Countries', ['class' => 'nav-link'.($currentModule == strtolower('countries') ? ' active' : '')]); ?>
-
+			<?= anchor(route_to('cityList'), '<svg class="bi me-2" width="16" height="16"><use xlink:href="#table"/></svg> '.lang('Cities.moduleTitle'), ['class' => 'nav-link'.($currentModule == strtolower('cities') ? ' active' : '')]); ?>
 		</li>
 		<li class="nav-item">
-			<?= anchor(route_to('cities'), '<i class="far fa-circle nav-icon"></i> Cities', ['class' => 'nav-link'.($currentModule == strtolower('cities') ? ' active' : '')]); ?>
-
+			<?= anchor(route_to('countryList'), '<svg class="bi me-2" width="16" height="16"><use xlink:href="#table"/></svg> '.lang('Countries.moduleTitle'), ['class' => 'nav-link'.($currentModule == strtolower('countries') ? ' active' : '')]); ?>
 		</li>
 		<li class="nav-item">
-			<?= anchor(route_to('people'), '<i class="far fa-circle nav-icon"></i> People', ['class' => 'nav-link'.($currentModule == strtolower('people') ? ' active' : '')]); ?>
-
+			<?= anchor(route_to('personList'), '<svg class="bi me-2" width="16" height="16"><use xlink:href="#table"/></svg> '.lang('People.moduleTitle'), ['class' => 'nav-link'.($currentModule == strtolower('people') ? ' active' : '')]); ?>
 		</li>
